*move all code to polygon.cpp and polygon.h *fixed bug wih 16 test
